do not remove file when extension is none

+++ ckstyle/doCssCompress.py
@@ -40,7 +40,7 @@ def compressFile(filePath, config = defaultConfig):
     else:
         path = os.path.realpath(filePath.split('.css')[0] + extension)
     if config.printFlag:
-        if os.path.exists(path):
+        if extension is not None and os.path.exists(path):
             os.remove(path)
         console.show(message)
     else:
